% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBOE.R
\name{DBOE}
\alias{DBOE}
\title{Database Object Explorer}
\description{
\code{DBOE} facilitates the navigation of SQL-based database engines by means of metadata. Engines tested to date include Microsoft SQL Server and MySQL.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{connection.list}}{Sets or returns a list of saved connections.  When providing a list, it should be names by database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBOE-new}{\code{DBOE$new()}}
\item \href{#method-DBOE-clone}{\code{DBOE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{\link[rlang]{dots_list}}: named DBI-compliant connection objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
