% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metadata.R
\name{dboe_lazy_table}
\alias{dboe_lazy_table}
\title{DBOE Lazy Table}
\usage{
dboe_lazy_table(env = new.env(parent = emptyenv()), name = NULL, meta = list())
}
\description{
\code{dboe_lazy_table} is a simple S7 class to represent a database table with attached metadata.
}
\section{Slots}{

\describe{
\item{\code{env}}{An environment containing the database connection and other relevant information.}

\item{\code{name}}{The qualified name of the table (e.g., \code{DBI::Id("schema", "table")}).}

\item{\code{tbl}}{A property that lazily evaluates to a \code{dplyr::tbl()} when accessed.}

\item{\code{meta}}{A list of metadata information for the table, structured as a list of column metadata.}
}}

\keyword{internal}
