% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metadata.R
\name{get.metadata}
\alias{get.metadata}
\title{Get Database Metadata}
\usage{
get.metadata(dboe, ...)
}
\arguments{
\item{dboe}{A \link{DBOE} object.}

\item{...}{(\code{\link[rlang]{dots_list}}) One or more symbols/names of connection environments
if they already exist in the object environment.

For each database connection, an environment is
created within \code{dboe} containing a \code{metamap} object with columns \code{database},
\code{tbl_name}, \code{col_name}, \code{schema_name}, \code{column_id}, and \code{data_type}.
Additionally, \code{\link{dboe_lazy_table}} objects are created for each table endpoint.}
}
\value{
Side effects only: warnings are thrown for non-existent connection environment names passed to \code{...}.
}
\description{
Retrieves metadata information for each database connection stored in a
\link{DBOE} object.
}
