% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.for.R
\name{\%look.for\%}
\alias{\%look.for\%}
\title{Look for a Database Metadata Reference}
\usage{
haystack \%look.for\% x
}
\arguments{
\item{haystack}{An object. Either a \code{metamap} object or a database
environment, for example \code{DBOE$database}, containing \code{metamap}.}

\item{x}{A character vector of regular expression patterns or exact names to
match against database object names.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with matching items, if any are found.
}
\description{
Searches a metadata environment or \code{metamap} object for the regular
expression patterns passed to \code{x}.
}
