% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.db_connection.R
\name{make.db_connection}
\alias{make.db_connection}
\title{Make a Database Connection}
\usage{
make.db_connection(..., drv = odbc::odbc(), default)
}
\arguments{
\item{...}{Named arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{drv}{A DBI-compliant driver object.}

\item{default}{An optional named list of default connection arguments.}
}
\value{
A DBI-compliant connection object with a \code{call} attribute storing the
connection expression.
}
\description{
Helper that wraps \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
