% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.virtual_database.R
\name{make.virtual_database}
\alias{make.virtual_database}
\title{Make a Virtual Database of Table Links}
\usage{
make.virtual_database(
  dboe = NULL,
  conn = NULL,
  ...,
  target_env,
  sch = "dbo",
  exclude = NA,
  lazy = TRUE
)
}
\arguments{
\item{dboe}{A \link{DBOE} object.}

\item{conn}{A symbol or character string naming a metadata environment
created after calling \code{\link[=get.metadata]{get.metadata()}}.}

\item{...}{\code{\link[rlang]{dots_list}}: Symbols, calls, \code{DBI::Id()} objects, or character strings for
which table pointers will be made. Elements containing \code{.} are treated as
qualified names, allowing cross-database endpoints to be referenced from a
single connection.
Nested lists will be flattened with names following the behavior of \code{\link[base]{unlist}}.}

\item{target_env}{An environment in which created objects should be stored.
If omitted, \code{dboe[[conn]]} is attached as \verb{dboe_<conn>}.}

\item{sch}{A symbol or character string giving the target schema to use when
selecting linked objects.}

\item{exclude}{A character vector of patterns or names to exclude from the
search.}

\item{lazy}{A boolean indicating whether to include \code{dboe_lazy_table} objects.
Arguments to the dots list with matching names in the connection environment are overriden.}
}
\value{
Side effects only. DBI-sourced \code{\link[dplyr:tbl]{dplyr::tbl()}} objects are assigned to
\code{target_env}.
}
\description{
Creates DBI-backed \code{\link[dplyr:tbl]{dplyr::tbl()}} objects in a target environment.
}
