% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBOE.R
\name{query_adhoc}
\alias{query_adhoc}
\title{Retrieve the contents of a query}
\usage{
query_adhoc(url, ...)
}
\arguments{
\item{url}{(string) The URL to call}

\item{...}{Arguments to be passed to \code{\link[keyring]{key_get}}}
}
\value{
The content of the URL
}
\description{
\code{query_adhoc} calls \code{\link[httr]{GET}}, and passes the credentials referenced by argument \code{cred}.  The response is converted to class 'character' before returning the output
}
