% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBOE.R
\name{DBOE}
\alias{DBOE}
\title{Database Object Explorer}
\description{
\code{DBOE} facilitates the navigation of SQL-based database engines by means of metadata. Engines tested to date include Microsoft SQL Server and MySQL.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{connection.list}}{Sets or returns a list of saved connections.  When providing a list, it should be names by database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBOE-new}{\code{DBOE$new()}}
\item \href{#method-DBOE-get.metadata}{\code{DBOE$get.metadata()}}
\item \href{#method-DBOE-make.virtual_database}{\code{DBOE$make.virtual_database()}}
\item \href{#method-DBOE-clone}{\code{DBOE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The class object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-get.metadata"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-get.metadata}{}}}
\subsection{Method \code{get.metadata()}}{
\code{$get.metadata()} retrieves metadata information for the database pointed to by argument \code{conns}. Once metadata has been retrieved, metadata can be accessed for tables, views, and stored procedures using the following access method: \code{<DBOE obj>$<database name>$<table/view/proc name>}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$get.metadata(..., chatty = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{\link[rlang]{dots_list}}: one or more DBI/ODBC connection objects.}

\item{\code{chatty}}{(logical) When \code{TRUE}, additional execution messages are sent to the console}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The class object, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-make.virtual_database"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-make.virtual_database}{}}}
\subsection{Method \code{make.virtual_database()}}{
\code{$make.virtual_database} creates a set of \code{\link[dplyr]{tbl}} objects in an environment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$make.virtual_database(conn, target_env = rlang::caller_env(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{The name of a metadata environment (created after calling \code{$get.metadata() })}

\item{\code{target_env}}{The environment object where created objects should be stored}

\item{\code{...}}{Names or patterns of objects to retrieve to link from source.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An assignable environment object with \code{DBI}-sourced \code{\link[dplyr]{tbl}}s
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBOE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
