% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.metadata.R
\name{get.metadata}
\alias{get.metadata}
\title{Get Database Metadata}
\usage{
get.metadata(dboe)
}
\arguments{
\item{dboe}{A \code{\link[DBOE]{DBOE}} object}
}
\value{
\itemize{
\item{An object named \code{metamap} is created in \code{<DBOE obj>$<database name>} having the following columns: \code{database}, \code{tbl_name}, \code{col_name}, \code{schema_name}, \code{column_id}, \code{data_type}.}
\item{For each table name in \code{metamap}, an object is created in \code{<DBOE obj>$<database name>} containing column metadata. This is essentially the same data, but allows for easier navigation using R's accessor function.}
}
}
\description{
\code{get.metadata()} retrieves metadata information for the database pointed to by argument \code{conns}. Once metadata has been retrieved, metadata can be accessed for tables, views, and stored procedures using the following access method: \code{<DBOE obj>$<database name>$<table/view/proc name>}
}
