% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBOE.R
\name{\%look.for\%}
\alias{\%look.for\%}
\title{Look for a Database Object Names}
\usage{
i \%look.for\% x
}
\arguments{
\item{i}{(object) A data.frame, data.table, or coercible containing DBMS object names}

\item{x}{(string[]) A vector of REGEX patterns or exact names to use for matching against database object names}
}
\value{
A \code{\link[data.table]{data.table}} object with the items that were found, if any
}
\description{
Look for a Database Object Names
}
