% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.db_connection.R
\name{make.db_connection}
\alias{make.db_connection}
\title{Make an ODBC MSSQL Database Connection}
\usage{
make.db_connection(..., drv = odbc::odbc(), default)
}
\arguments{
\item{...}{Valid values passed to \code{\link[DBI]{dbConnect}}.}

\item{drv}{A DBI-compliant driver}
}
\value{
A \code{\link[DBI]{DBI}}-compliant connection object.
}
\description{
This is a helper function that wraps \code{\link[DBI]{dbConnect}}
}
